Create Product & Transaction Repository & Service

+++ src/main/java/id/ac/ui/cs/advprog/gametime/transaction/model/Builder/ProductBuilder.java
@@ -8,7 +8,7 @@ public class ProductBuilder {
     private String name;
     private String description;
     private String category;
-    private double price;
+    private int price;
     private double rating;
 
     public ProductBuilder seller(User seller) {
@@ -47,7 +47,7 @@ public class ProductBuilder {
         return this;
     }
 
-    public ProductBuilder price(double price) {
+    public ProductBuilder price(int price) {
         if (price <= 0) {
             throw new IllegalArgumentException("Price must be more than 0");
         }
